implemented structure to support upcoming features
added bin_to_program and compile procedures in main which at the moment
only have a basic implementation. Started using these procedures for the
options.bin and options.compile code paths. The only thing missing is a
correct implementation of these procedures for both of these features to
be completely implemented