implemented debugger partially
Command parsing and repl interactions are almost done (a lot of refinements
still need to be made but the rough structure is done and it works). What is
missing is the actual implementation for each of the commands.

On a sidenote: the library for repl interactions (nim-noise) has a bug which
has been fixed on github but not in the newest release as far as i can tell.